refac: factor out db write methods
into a mixin mdb_update
